+++ sdk/src/HyperTrack/data_types/OrderStatus.ts
@@ -1,0 +1,11 @@
+export type OrderStatus =
+  | {
+      type: 'orderStatusClockIn';
+    }
+  | {
+      type: 'orderStatusClockOut';
+    }
+  | {
+      type: 'orderStatusCustom';
+      value: string;
+    };