+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [13.3.0] - 2024-04-08
 
+### Added
+
+- New `addGeotag` method that have `orderHandle` and `orderStatus` parameters. You can use this API when users need to clock in/out of work in your app to honor their work time (see [Clock in/out Tagging](https://hypertrack.com/docs/clock-inout-tracking#add-clock-inout-events-to-a-shift-timeline) guide for more info)
+
 ### Changed
 
 - Updated HyperTrack SDK iOS to [5.5.0](https://github.com/hypertrack/sdk-ios/releases/tag/5.5.0)